Render the bottom bar and animate it
