+++ packages/domain-search/src/components/domains-mini-cart/actions.tsx
@@ -1,16 +1,18 @@
 import { __experimentalHStack as HStack, Button } from '@wordpress/components';
+import { useI18n } from '@wordpress/react-i18n';
 import { useDomainSearch } from '../domain-search';
 
 export const DomainsMiniCartActions = () => {
 	const { onContinue, openFullCart } = useDomainSearch();
+	const { __ } = useI18n();
 
 	return (
 		<HStack spacing={ 2 }>
 			<Button variant="tertiary" onClick={ () => openFullCart() }>
-				View cart
+				{ __( 'View cart' ) }
 			</Button>
 			<Button variant="primary" onClick={ () => onContinue() }>
-				Continue
+				{ __( 'Continue' ) }
 			</Button>
 		</HStack>
 	);
