+++ packages/domain-search/src/index.ts
@@ -1,5 +1,5 @@
 export { DomainSearch, type DomainSearchCart } from './components/domain-search';
 export { DomainSearchControls } from './components/DomainSearchControls/DomainSearchControls';
-export { DomainsMiniCart } from './components/DomainsMiniCart/DomainsMiniCart';
+export { DomainsMiniCart } from './components/domains-mini-cart';
 export { DomainsFullCart } from './components/domains-full-cart';
 export * as DomainSuggestions from './components/DomainSuggestions';
